Create basic two basic test cases and a way to visualize them
A test case is just a correctly formatted query. If a formatted version of a test case is equal to the test case is equal to the test case, then the test is passed. This assumes that the first step of the formattign will remove all spaces and newlines so that we are sure that actual formatting will take place. Otherwise doing no formatting at all would guarantee that all test cases were always passed