[Fix #58] Refactor all files to use require imports
Previously, the bundler did not support relative paths, so I could
either use require with absolute paths, which would be fine with the
bundler, but could not easily(?) be utilized by Node, or I could use
include imports, which would not cause any conflicts with regular
require imports, but could also not be run natively by Node.

This commit updates the bundler to support relative imports as used by
require. This means that all IDE plugins will work as expected, and
all the code will be valid JavaScript, even before being compiled,
which could be an advantage.

Among other things, this will make it easier to extract the Gremlint
formatQuery function out to a separate library and still use it in the
gremlint.com website.

It also removes the need for the custom node run script, which used to
be a workaround for being able to run files which did not use valid
require imports.