[Fix #71] Make indentation work for top-level non-gremlin code
For simplicity, the code is formatted with the assumption that there
is no initial indentation, and that the max line length is the user-
specified max line length minus the user-specified indentation.

After the query is formatted, the user-specified indentation is added,
but only to non-empty lines, because we don't want empty lines with
hidden spaces.